backend/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.3.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
boto3==1.39.9
botocore==1.39.9
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.1
cryptography==45.0.5
cv-3==1.2.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
Django==3.1.3
django-filter==2.4.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
filterpy==1.4.5
Flask==3.1.1
Flask-PyMongo==3.0.1
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hub-sdk==0.0.20
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
labelImg==1.8.6
labeling==0.1.14
lap==0.5.12
lazy_loader==0.4
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
narwhals==1.43.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
omegaconf==2.3.0
openai==1.97.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pfzy==0.3.4
pillow==11.1.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.13.2
PyNaCl==1.5.0
pyparsing==3.2.1
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyttsx==1.1
pyttsx3==2.98
pytz==2020.4
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.25.1
runpod==1.7.13
s3transfer==0.13.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.33.0
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.4.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.46.0
supervision==0.25.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.13
tk==0.1.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
tqdm-loggable==0.2
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
ultralytics==8.3.97
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
